Convert location module
